requirements/prod.txt
celery==5.4.0
Django==4.2.16
django-betterforms==2.0.0
django-celery-results==2.5.1
django-ipware==7.0.1
django-ratelimit==4.1.0
django-simple-captcha==0.6.0
django_redis==5.4.0
gevent==24.11.1
Pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
redis==5.2.1
gunicorn==23.0.0
